Move preferences to preferences namespace
