+++ app/src/main/java/org/jellyfin/androidtv/TvApp.java
@@ -25,10 +25,10 @@ import org.jellyfin.androidtv.playback.MediaManager;
 import org.jellyfin.androidtv.playback.PlaybackController;
 import org.jellyfin.androidtv.playback.PlaybackManager;
 import org.jellyfin.androidtv.playback.PlaybackOverlayActivity;
+import org.jellyfin.androidtv.preferences.SystemPreferences;
+import org.jellyfin.androidtv.preferences.UserPreferences;
 import org.jellyfin.androidtv.search.SearchActivity;
 import org.jellyfin.androidtv.util.DeviceUtils;
-import org.jellyfin.androidtv.util.SystemPreferences;
-import org.jellyfin.androidtv.util.UserPreferences;
 import org.jellyfin.androidtv.util.Utils;
 import org.jellyfin.apiclient.interaction.ApiClient;
 import org.jellyfin.apiclient.interaction.EmptyResponse;
