+++ app/src/main/java/org/jellyfin/androidtv/livetv/GuideFilters.java
@@ -1,7 +1,7 @@
 package org.jellyfin.androidtv.livetv;
 
 import org.jellyfin.androidtv.TvApp;
-import org.jellyfin.androidtv.util.SystemPreferences;
+import org.jellyfin.androidtv.preferences.SystemPreferences;
 import org.jellyfin.androidtv.util.Utils;
 import org.jellyfin.apiclient.model.dto.BaseItemDto;
