+++ app/src/main/java/org/jellyfin/androidtv/livetv/TvManager.java
@@ -10,8 +10,8 @@ import org.jellyfin.androidtv.R;
 import org.jellyfin.androidtv.TvApp;
 import org.jellyfin.androidtv.itemhandling.ItemRowAdapter;
 import org.jellyfin.androidtv.model.LiveTvPrefs;
+import org.jellyfin.androidtv.preferences.SystemPreferences;
 import org.jellyfin.androidtv.ui.ProgramGridCell;
-import org.jellyfin.androidtv.util.SystemPreferences;
 import org.jellyfin.androidtv.util.TimeUtils;
 import org.jellyfin.androidtv.util.Utils;
 import org.jellyfin.apiclient.interaction.EmptyResponse;
