+++ app/src/main/java/org/jellyfin/androidtv/playback/ExternalPlayerActivity.java
@@ -13,8 +13,8 @@ import org.jellyfin.androidtv.TvApp;
 import org.jellyfin.androidtv.model.compat.PlaybackException;
 import org.jellyfin.androidtv.model.compat.StreamInfo;
 import org.jellyfin.androidtv.model.compat.VideoOptions;
+import org.jellyfin.androidtv.preferences.UserPreferences;
 import org.jellyfin.androidtv.util.ProfileHelper;
-import org.jellyfin.androidtv.util.UserPreferences;
 import org.jellyfin.androidtv.util.Utils;
 import org.jellyfin.androidtv.util.apiclient.ReportingHelper;
 import org.jellyfin.apiclient.interaction.Response;
