+++ app/src/main/java/org/jellyfin/androidtv/preferences/SharedPreferenceStore.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.androidtv.util
+package org.jellyfin.androidtv.preferences
 
 import android.content.SharedPreferences
 import org.jellyfin.androidtv.TvApp
